switch over to github actions for CI (#153)
This ditches travis and appveyor in favor of github actions.

We're going for a fresh build of the target php-src branch, so that when the build is run by cron, we really do have *continuous* integration and I get warning early when the build breaks because of changes in php-src. It also means we don't have to rely on anyone packaging anything and we have complete control over the build.

TODO: there's no memcheck build yet, and no mac build.