Use same technique as for cuf(a)
uopz already implements its own `call_user_func(_array)`, so we do the
same with `exit`.  While the former has different reasons (compiler
might optimize these calls away), the use case is similar enough to not
use a different mechanism.  This is also a bit cleaner, and may avoid
crashes[1], at the cost of needing to maintain a copy of the exit
function implementation.

[1] <https://github.com/krakjoe/uopz/commit/e902fb2bdce6c59e1fa5a221c98a8cb73ada4ad9>